BREAKING CHANGE: replace `children` with `actions`
IMO there is no advantage to using the children feature of
React over passing in an prop that is far more clear as to
what it is. Children could end up anywhere. Naming it actions
it's clear that it's for actions, not anything. You lose
the nested html JSX thing, but boo hoo, it doesn't even make any sense for this component
anyway.